feat: create order

+++ RestaurantManagementServer/src/main/java/dao/CategoryDAO.java
@@ -3,7 +3,7 @@ package dao;
 import jakarta.persistence.EntityManager;
 import model.CategoryEntity;
 
-public class CategoryDAO extends GenericDAO<CategoryEntity, Integer> {
+public class CategoryDAO extends GenericDAO<CategoryEntity, Long> {
     public CategoryDAO() {
         super(CategoryEntity.class);
     }
