+++ RestaurantManagementServer/src/main/java/dao/FloorDAO.java
@@ -3,7 +3,7 @@ package dao;
 import jakarta.persistence.EntityManager;
 import model.FloorEntity;
 
-public class FloorDAO extends GenericDAO<FloorEntity, Integer> {
+public class FloorDAO extends GenericDAO<FloorEntity, Long> {
     public FloorDAO() {
         super(FloorEntity.class);
     }
