+++ RestaurantManagementServer/src/main/java/dao/RoleDAO.java
@@ -3,7 +3,7 @@ package dao;
 import jakarta.persistence.EntityManager;
 import model.RoleEntity;
 
-public class RoleDAO extends GenericDAO<RoleEntity, Integer> {
+public class RoleDAO extends GenericDAO<RoleEntity, Long> {
     public RoleDAO() {
         super(RoleEntity.class);
     }
